CORTX-29883-Cluster Start
